// frontend-admin/src/api/generated/model/receiptListItemDtoPagedResult.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */
import type { ReceiptListItemDto } from './receiptListItemDto';

export interface ReceiptListItemDtoPagedResult {
  /** @nullable */
  items?: ReceiptListItemDto[] | null;
  page?: number;
  pageSize?: number;
  totalCount?: number;
  totalPages?: number;
}